requirements.txt
google_api_python_client==2.126.0
Pillow==9.5.0
Requests==2.32.3
streamlit==1.36.0
streamlit_lottie==0.0.5
